In-progress commit on TLP rework.
The largest change is the code for temporary SQLite databases. It
provides... a possibly overwrought abstraction for setting up and
writing to the DBs.

This cleaned up the random SQL code woven throughout the temporary DB
space. Or, it created a new (but tested) mess of SQLite code. I think it
added a bit of safety, because of unchecked strings in the code, it now
has structures and values that get (some) checking before SQL is
executed.

The kill channels have been replaced with a broker, which allows
processes to subscribe to kill messages. It still gets threaded through
the TLP, but the PAR gets burried in the broker.

The reset channels are similarly now a broker. This, again, means that
processes that should do something "on reset" now subscribe. (I have a
suspicion we could reduce the number of processes listening for a reset
down to one...)

Some renaming of processes...